#362 Fix bug in media title
Fix System.AccessViolationException when playing files with a title in the metadata.
Title was being treated as a string but represents the index of the "Title" being played.

+++ src/Vlc.DotNet.Core.Interops/Signatures/libvlc_events.h/libvlc_event_t.cs
@@ -163,9 +163,8 @@ namespace Vlc.DotNet.Core.Interops.Signatures
 
     [StructLayout(LayoutKind.Sequential)]
     public struct MediaPlayerTitleChanged
-    {
-        //todo : original was int : Check if ok
-        public IntPtr NewTitle;
+    {        
+        public int NewTitle;
     }
 
     [StructLayout(LayoutKind.Sequential)]
